refactor: `Email Message` to create and submit email messages
