+++ mail/mail/doctype/email_message/email_message.js
@@ -3,29 +3,87 @@
 
 frappe.ui.form.on('Email Message', {
 	refresh(frm) {
-		if (!frm.doc.__islocal && !frm.doc.destroyed) {
-			frm.trigger('add_reply_buttons')
-			frm.trigger('add_forward_button')
-			frm.trigger('add_mark_buttons')
-			frm.trigger('add_move_buttons')
-			frm.trigger('add_actions')
+		frm.disable_save()
+
+		if (!frm.doc.destroyed) {
+			if (!frm.doc.__islocal) {
+				if (!frm.doc.draft) {
+					frm.trigger('add_reply_buttons')
+					frm.trigger('add_forward_button')
+					frm.trigger('add_mark_buttons')
+					frm.trigger('add_move_buttons')
+				}
+
+				frm.trigger('add_actions')
+			}
+
+			if (frm.doc.draft) {
+				frm.trigger('add_draft_button')
+				frm.trigger('add_send_button')
+			}
 		}
 	},
 
 	add_reply_buttons(frm) {
-		frm.add_custom_button(__('Reply'), () => frm.trigger('reply'))
-		frm.add_custom_button(__('Reply All'), () => frm.trigger('reply_all'))
+		frm.add_custom_button(__('Reply'), () => {
+			frappe.model.open_mapped_doc({
+				method: 'mail.mail.doctype.email_message.email_message.reply',
+				frm: frm,
+				freeze: true,
+				freeze_message: __('Loading...'),
+			})
+		})
+
+		frm.add_custom_button(__('Reply All'), () => {
+			frappe.model.open_mapped_doc({
+				method: 'mail.mail.doctype.email_message.email_message.reply_all',
+				frm: frm,
+				freeze: true,
+				freeze_message: __('Loading...'),
+			})
+		})
 	},
 
 	add_forward_button(frm) {
-		frm.add_custom_button(__('Forward'), () => frm.trigger('forward'))
+		frm.add_custom_button(__('Forward'), () => {
+			frappe.model.open_mapped_doc({
+				method: 'mail.mail.doctype.email_message.email_message.forward',
+				frm: frm,
+				freeze: true,
+				freeze_message: __('Loading...'),
+			})
+		})
 	},
 
 	add_mark_buttons(frm) {
 		if (frm.doc.seen) {
-			frm.add_custom_button(__('Mark as Unread'), () => frm.trigger('mark_as_unseen'))
+			frm.add_custom_button(__('Mark as Unread'), () => {
+				frappe.call({
+					doc: frm.doc,
+					method: 'mark_as_unseen',
+					freeze: true,
+					freeze_message: __('Marking as Unread...'),
+					callback: (r) => {
+						if (!r.exc) {
+							frm.refresh()
+						}
+					},
+				})
+			})
 		} else {
-			frm.add_custom_button(__('Mark as Read'), () => frm.trigger('mark_as_seen'))
+			frm.add_custom_button(__('Mark as Read'), () => {
+				frappe.call({
+					doc: frm.doc,
+					method: 'mark_as_seen',
+					freeze: true,
+					freeze_message: __('Marking as Read...'),
+					callback: (r) => {
+						if (!r.exc) {
+							frm.refresh()
+						}
+					},
+				})
+			})
 		}
 	},
 
@@ -55,66 +113,72 @@ frappe.ui.form.on('Email Message', {
 	},
 
 	add_actions(frm) {
-		const add_button = (label, trigger) => {
-			frm.add_custom_button(__(label), () => frm.trigger(trigger), __('Actions'))
+		if (frm.doc.has_attachment) {
+			frm.add_custom_button(
+				__('Load Attachments'),
+				() => {
+					frappe.call({
+						doc: frm.doc,
+						method: 'preload_attachments_to_cache',
+						args: {
+							include_inline: true,
+							include_regular: true,
+						},
+						freeze: true,
+						freeze_message: __('Loading Attachments...'),
+						callback: (r) => {
+							if (!r.exc) {
+								frm.refresh()
+							}
+						},
+					})
+				},
+				__('Actions'),
+			)
 		}
 
-		if (frm.doc.has_attachment) add_button('Load Attachments', 'load_attachments')
-		if (!frm.doc.message) add_button('Load MIME Message', 'load_mime_message')
+		if (!frm.doc.message) {
+			frm.add_custom_button(
+				__('Load MIME Message'),
+				() => {
+					frappe.call({
+						doc: frm.doc,
+						method: 'get_mime_message',
+						freeze: true,
+						freeze_message: __('Loading MIME Message...'),
+						callback: (r) => {
+							if (!r.exc) {
+								frm.refresh()
+							}
+						},
+					})
+				},
+				__('Actions'),
+			)
+		}
 	},
 
-	reply(frm) {
-		frappe.model.open_mapped_doc({
-			method: 'mail.mail.doctype.email_message.email_message.reply',
-			frm: frm,
-			freeze: true,
-			freeze_message: __('Loading...'),
+	add_draft_button(frm) {
+		frm.page.set_secondary_action(__('Save as Draft'), () => {
+			frappe.call({
+				doc: frm.doc,
+				method: 'save_draft',
+				freeze: true,
+				freeze_message: __('Saving Draft...'),
+				callback: (r) => {
+					if (!r.exc) {
+						if (r.message) {
+							frappe.set_route('Form', 'Email Message', r.message)
+						}
+					}
+				},
+			})
 		})
 	},
 
-	reply_all(frm) {
-		frappe.model.open_mapped_doc({
-			method: 'mail.mail.doctype.email_message.email_message.reply_all',
-			frm: frm,
-			freeze: true,
-			freeze_message: __('Loading...'),
-		})
-	},
-
-	forward(frm) {
-		frappe.model.open_mapped_doc({
-			method: 'mail.mail.doctype.email_message.email_message.forward',
-			frm: frm,
-			freeze: true,
-			freeze_message: __('Loading...'),
-		})
-	},
-
-	mark_as_unseen(frm) {
-		frappe.call({
-			doc: frm.doc,
-			method: 'mark_as_unseen',
-			freeze: true,
-			freeze_message: __('Marking as Unread...'),
-			callback: (r) => {
-				if (!r.exc) {
-					frm.refresh()
-				}
-			},
-		})
-	},
-
-	mark_as_seen(frm) {
-		frappe.call({
-			doc: frm.doc,
-			method: 'mark_as_seen',
-			freeze: true,
-			freeze_message: __('Marking as Read...'),
-			callback: (r) => {
-				if (!r.exc) {
-					frm.refresh()
-				}
-			},
+	add_send_button(frm) {
+		frm.page.set_primary_action(__('Send'), () => {
+			frm.save()
 		})
 	},
 
@@ -127,34 +191,6 @@ frappe.ui.form.on('Email Message', {
 			args: {
 				mailbox_role: mailbox_role,
 			},
-			callback: (r) => {
-				if (!r.exc) {
-					frm.refresh()
-				}
-			},
-		})
-	},
-
-	load_attachments(frm) {
-		frappe.call({
-			doc: frm.doc,
-			method: 'preload_attachments_to_cache',
-			freeze: true,
-			freeze_message: __('Loading Attachments...'),
-			callback: (r) => {
-				if (!r.exc) {
-					frm.refresh()
-				}
-			},
-		})
-	},
-
-	load_mime_message(frm) {
-		frappe.call({
-			doc: frm.doc,
-			method: 'get_mime_message',
-			freeze: true,
-			freeze_message: __('Loading MIME Message...'),
 			callback: (r) => {
 				if (!r.exc) {
 					frm.refresh()
