+++ mail/mail/doctype/email_message_part/email_message_part.json
@@ -13,6 +13,7 @@
   "disposition",
   "language",
   "location",
+  "file_url",
   "column_break_xvpk",
   "type",
   "blob_id",
@@ -27,24 +28,28 @@
   {
    "fieldname": "blob_id",
    "fieldtype": "Data",
-   "label": "Blob ID"
+   "label": "Blob ID",
+   "read_only": 1
   },
   {
    "fieldname": "size",
    "fieldtype": "Data",
    "in_list_view": 1,
-   "label": "Size (Bytes)"
+   "label": "Size (Bytes)",
+   "read_only": 1
   },
   {
    "fieldname": "type",
    "fieldtype": "Data",
    "in_list_view": 1,
-   "label": "Type"
+   "label": "Type",
+   "read_only": 1
   },
   {
    "fieldname": "charset",
    "fieldtype": "Data",
-   "label": "Charset"
+   "label": "Charset",
+   "read_only": 1
   },
   {
    "fieldname": "disposition",
@@ -54,6 +59,7 @@
    "options": "\ninline\nattachment"
   },
   {
+   "depends_on": "eval: doc.disposition == \"inline\"",
    "fieldname": "cid",
    "fieldtype": "Data",
    "label": "Content ID"
@@ -69,30 +75,38 @@
   {
    "fieldname": "part_id",
    "fieldtype": "Data",
-   "label": "Part ID"
+   "label": "Part ID",
+   "read_only": 1
   },
   {
    "fieldname": "language",
    "fieldtype": "Data",
-   "label": "Language"
+   "label": "Language",
+   "read_only": 1
   },
   {
    "fieldname": "location",
    "fieldtype": "Data",
-   "label": "Location"
+   "label": "Location",
+   "read_only": 1
   },
   {
    "fieldname": "filename",
    "fieldtype": "Data",
    "in_list_view": 1,
    "label": "File Name"
+  },
+  {
+   "fieldname": "file_url",
+   "fieldtype": "Attach",
+   "label": "File"
   }
  ],
  "grid_page_length": 50,
  "index_web_pages_for_search": 1,
  "istable": 1,
  "links": [],
- "modified": "2025-05-14 18:27:56.146189",
+ "modified": "2025-05-15 16:34:10.700371",
  "modified_by": "Administrator",
  "module": "Mail",
  "name": "Email Message Part",
