--- mail/mail/doctype/mail_queue_attachment/mail_queue_attachment.json
@@ -1,106 +1,0 @@
-{
- "actions": [],
- "allow_rename": 1,
- "creation": "2025-05-14 09:48:56.839164",
- "doctype": "DocType",
- "editable_grid": 1,
- "engine": "InnoDB",
- "field_order": [
-  "section_break_zm8h",
-  "filename",
-  "disposition",
-  "file_url",
-  "column_break_irmn",
-  "type",
-  "blob_id",
-  "cid",
-  "size",
-  "section_break_o2zj",
-  "preview"
- ],
- "fields": [
-  {
-   "fieldname": "section_break_zm8h",
-   "fieldtype": "Section Break"
-  },
-  {
-   "fieldname": "type",
-   "fieldtype": "Data",
-   "in_list_view": 1,
-   "label": "Type",
-   "no_copy": 1,
-   "read_only": 1
-  },
-  {
-   "fieldname": "disposition",
-   "fieldtype": "Select",
-   "in_list_view": 1,
-   "label": "Disposition",
-   "options": "\ninline\nattachment",
-   "reqd": 1
-  },
-  {
-   "fieldname": "column_break_irmn",
-   "fieldtype": "Column Break"
-  },
-  {
-   "fieldname": "blob_id",
-   "fieldtype": "Data",
-   "label": "Blob ID",
-   "no_copy": 1,
-   "read_only": 1
-  },
-  {
-   "depends_on": "eval: doc.disposition == \"inline\"",
-   "fieldname": "cid",
-   "fieldtype": "Data",
-   "label": "Content ID"
-  },
-  {
-   "fieldname": "size",
-   "fieldtype": "Data",
-   "in_list_view": 1,
-   "label": "Size (Bytes)",
-   "no_copy": 1,
-   "read_only": 1
-  },
-  {
-   "depends_on": "eval: doc.file_url",
-   "fieldname": "preview",
-   "fieldtype": "Image",
-   "label": "Preview",
-   "options": "file_url"
-  },
-  {
-   "fieldname": "section_break_o2zj",
-   "fieldtype": "Section Break"
-  },
-  {
-   "fieldname": "filename",
-   "fieldtype": "Data",
-   "in_list_view": 1,
-   "label": "File Name"
-  },
-  {
-   "fieldname": "file_url",
-   "fieldtype": "Attach",
-   "in_list_view": 1,
-   "label": "File",
-   "reqd": 1
-  }
- ],
- "grid_page_length": 50,
- "index_web_pages_for_search": 1,
- "istable": 1,
- "links": [],
- "modified": "2025-05-14 18:22:47.983290",
- "modified_by": "Administrator",
- "module": "Mail",
- "name": "Mail Queue Attachment",
- "owner": "Administrator",
- "permissions": [],
- "row_format": "Dynamic",
- "sort_field": "creation",
- "sort_order": "DESC",
- "states": []
-}
