--- mail/mail/doctype/mail_queue_recipient/mail_queue_recipient.json
@@ -1,61 +1,0 @@
-{
- "actions": [],
- "creation": "2025-05-08 15:11:31.353620",
- "doctype": "DocType",
- "editable_grid": 1,
- "engine": "InnoDB",
- "field_order": [
-  "section_break_sxco",
-  "type",
-  "column_break_9abe",
-  "display_name",
-  "email"
- ],
- "fields": [
-  {
-   "fieldname": "section_break_sxco",
-   "fieldtype": "Section Break"
-  },
-  {
-   "default": "To",
-   "fieldname": "type",
-   "fieldtype": "Select",
-   "in_list_view": 1,
-   "label": "Type",
-   "options": "To\nCc\nBcc",
-   "reqd": 1
-  },
-  {
-   "fieldname": "email",
-   "fieldtype": "Data",
-   "in_list_view": 1,
-   "label": "Email",
-   "options": "Email",
-   "reqd": 1
-  },
-  {
-   "fieldname": "display_name",
-   "fieldtype": "Data",
-   "in_list_view": 1,
-   "label": "Display Name"
-  },
-  {
-   "fieldname": "column_break_9abe",
-   "fieldtype": "Column Break"
-  }
- ],
- "grid_page_length": 50,
- "index_web_pages_for_search": 1,
- "istable": 1,
- "links": [],
- "modified": "2025-05-08 18:26:46.847915",
- "modified_by": "Administrator",
- "module": "Mail",
- "name": "Mail Queue Recipient",
- "owner": "Administrator",
- "permissions": [],
- "row_format": "Dynamic",
- "sort_field": "creation",
- "sort_order": "DESC",
- "states": []
-}
