--- mail/mail/doctype/mail_queue_recipient/mail_queue_recipient.py
@@ -1,11 +1,0 @@
-# Copyright (c) 2025, Frappe Technologies Pvt. Ltd. and contributors
-# For license information, please see license.txt
-
-# import frappe
-from frappe.model.document import Document
-from uuid_utils import uuid7
-
-
-class MailQueueRecipient(Document):
-	def autoname(self) -> None:
-		self.name = str(uuid7())
