--- mail/mail/doctype/mail_queue_reply_to/mail_queue_reply_to.json
@@ -1,50 +1,0 @@
-{
- "actions": [],
- "creation": "2025-05-10 13:36:25.244980",
- "doctype": "DocType",
- "editable_grid": 1,
- "engine": "InnoDB",
- "field_order": [
-  "section_break_sm3z",
-  "display_name",
-  "column_break_ysao",
-  "email"
- ],
- "fields": [
-  {
-   "fieldname": "section_break_sm3z",
-   "fieldtype": "Section Break"
-  },
-  {
-   "fieldname": "display_name",
-   "fieldtype": "Data",
-   "in_list_view": 1,
-   "label": "Display Name"
-  },
-  {
-   "fieldname": "column_break_ysao",
-   "fieldtype": "Column Break"
-  },
-  {
-   "fieldname": "email",
-   "fieldtype": "Data",
-   "in_list_view": 1,
-   "label": "Email",
-   "reqd": 1
-  }
- ],
- "grid_page_length": 50,
- "index_web_pages_for_search": 1,
- "istable": 1,
- "links": [],
- "modified": "2025-05-10 13:40:15.594279",
- "modified_by": "Administrator",
- "module": "Mail",
- "name": "Mail Queue Reply To",
- "owner": "Administrator",
- "permissions": [],
- "row_format": "Dynamic",
- "sort_field": "creation",
- "sort_order": "DESC",
- "states": []
-}
